Update examples to use new DiskCacheCellImgFactory features
Additional options can now be given when calling create() to apply for
individual images. In particular, the option initializeCellsAsDirty(true)
can is used to make sure that results of expensive operations are written
to disk cache instead of recomputed. Before, we used the option
dirtyAccesses(false) for the same purpose. But with that, evicted cells
were *always* written (not only initially or when they had changed).